requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.31.0

# S3 Storage
aioboto3>=12.2.0
boto3>=1.29.0

# HTTP and HTML parsing
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=5.1.0
selectolax>=0.3.17

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# UI Templates
jinja2>=3.1.0
bleach>=6.0.0
jdatetime>=4.1.0

# Browser automation for IRNA
playwright>=1.40.0

